Add support for setting the Exercise ID.
When set, all outgoing PDUs will be given the configured exercise ID. On
the incoming side, PDUs without the specified exercise ID will be
discarded.

Updates pushed through to the disruptor and duplicator so that the
exercise ID can be set on the command line for their respective purposes
(send and record).