Fix spurious wait when bundling is enabled causing massive slow down.
When submit() is called in Bundler, it will arm the Sender thread which
will flush the queue on a timeout if there is no flush call made.

Added a condition to let the submit method wait until the Sender thread
confirms that it is armed. If we don't, submit() may be called many times
before the thread is interrupted. Possibly enough to hit the size limit,
causing us to signal the flush condition (before the Sender has had a
chance to arm).

We then wait for the return condition, yielding and giving the Sender a
chance to run. The first thing it does it arm, and then wait for the
flush signal (which is long gone by now). Without that, it waits for
max_wait_time.

The added condition lets the submit() method ensure that the Sender is
armed before it returns, solving the spurious wait.

Fix #8 - The drops are being caused slow consumption in the reflector.
If the sender is shut down, then it stops processing and loses the
queue, thus causing pdus to not be sent.