Fix problem with wrong broadcast address being chosen on MacOS
MacOS was returning an IPv6 address when LOOPBACK was set and the
original "Win|Mac" logic was in place. New method fixes problem for
choosing an IPv6 address (will now pick v4), however, it seems that like
Linux, we need to bind to the .255 address even for loopback on MacOS as
well as Linux.

I could have sworn I already tested this and it was the other way, but I
may have just been staying up too late. This has fixed a problem that
was preventing a local distributor from receiving traffic from a local
disruptor on the same machine.