Fix bug in EntityType.equals() and add hashCode for EntityType and ID
There was a stupid, stupid, stupid bug in EntityType.equals(). It would
compare the two object and if all fields were equal, it would return ...
false. WOT. Yes. Sigh.

This in turn caused massive object allocation when analyzing dumped
files as the compare option was used in determining whether we had
already stored a record for a particular type, which failed given the
above, and caused another record to be created and added. Combine that
with a sample dataset that is some 8.5M+ recorded PDUs big and I am
surprised that things ran for as long as they did before the OOM
problems started popping up.

Working: #11