Refactor Disassembler to be cleaner and ready for extension.
Now that I've dug a little way into this I've seen all the unnecessary
complexity that was added. This update refactors the Disassembler to
remove much of that while also making it ready for extension via the
creation of other analyzers.

The JSON output is currently broken and it is only looking a ESPDUs for
the time being, but those shall be fixed in next commits (sleep time).

Working #11