Cause BROADCAST to resolve properly for LOOPBACK. Log all NICs.
Due to operating-system specific things previously, we were returning
127.0.0.1 if you asked for the broadcast address with the loopback NIC.
This certainly does not work on windows. I _think_ it does on Mac, which
is why it would be there (need to confirm this doesn't break that).

Now we will ask the NIC for it's broadcast address and use that instead.
We will also log more NIC information at the start of each Disco app.

Fix: #21