Add support to dislocator for a "system test" function.
The system test function puts the dislocator into file replay mode, sets
the file to etc/dislocator-test.session and sets the tracking entity to
"Test" (unless the user has explicitly set the tracking entity name).

A session with a Predator UAV using the marking "Test" is included.
It is doing a figure 8 near Rockhampton in Queensland, Australia.

The user can replace the session file with their own. If they specify a
target marking that will also be honoured.