Change the dislocator NMEA server to generate updates on an interval
This cleans up the NMEA server a bit and changes the way that it works
from a update-in/update-out style (where each DIS update triggers a NMEA
message) to a heartbeat style. The server takes updates as they come,
but will only drop updates out to connected clients on an interval. This
helps manage the update rate, as clients on the other end typically
don't expect them to fly as thick and as fast as DIS might generate
them.