Add PDU processing flags and update Duplicator to never parse PDUs.
Flags can now be set statically on DiscoConfiguration. These help
control how PDUs are parsed. These include:

  - Strict: If set, PDU parsing will throw exception if it can't
            understand the contents of a PDU (e.g. the value of a field
            or enum is not valid). If off (the default), when an unknown
            field is encountered, a logical fallback is used.

  - Unparsed: If the PDU type is not supported, normally an exception is
              thrown. This mode instead wraps it in an UnparsedPdu type,
              which just body as an opaque byte[].

  - UnparsedExclusive: Always use UnparsedPdus; never attempt to parse
                       a PDU into a class that can read it.

Unparsed PDUs are useful for applications that only want to move PDUs
around and don't care about reading any data beyond the header.

This commit sets the Duplicator to use Unparsed PDUs all the time. It
doesn't care about the body of the PDU, it just wants to write them to
disk and then back to the network. This means the Duplicator can now
support capturing and replaying PDUs that Disco doesn't currently
support.

A command line argument has been added to enable STRICT mode for the
duplicator, which will cause it to discard PDUs that it can't
understand.

NOTE: Flags are static on DiscoConfiguration, so if you have a runtime
      with mulitple OpsCenter's, that may cause issues.