Refactor RprConnection and add basic EmitterStore to Disco App Framework
Significant refactor to the RprConnection to make it less tightly
coupled and full of boilerplate code. Includes very, very basic MessageBus
for the purposes of shunting internal messages around.

Added an EmitterStore to the Disco Application framework to make storage
and retrieval of emitter system data easier.

Added basic skeleton (unfinished) for Emitter DIS/RPR translation to the
RprConnection. This requires completion.