Add support for DIS<>HLA Emitter Translations through RprConnection.
Initial support added for translating between DIS and HLA/RPR types
through the RprConnection added. PDUs are generated for emitter system
and emitter beam changes on the RPR side, and objects are registered and
updated on the HLA side for any PDUs received.

NOTE: Currently has a limitation where references to objects that an
object is attached to can only work for objects generated on the same
side. For example, a RadioTransmitter can only successfully reference
its "host entity" if that entity is also created by the RprConnection.