Fix problem with CNR transmitters not traversing DIS/HLA bridge.
We test an object to see if it has been loaded propertly when we receive
a reflection from the HLA. This is so that we don't send out pdus with
partial information to network before we have received a "full reflect"
(or as much as we can expect of one). Unfortunately we were testing
Transmitters and Entities to see if their IDs had 0-0-0 (which is what
CNR uses for unattached radios). CNR is probably wrong in this instance,
but the check isn't good either. We just want to know if we're received
an update for all "necessary" fields.

As such, I've added a method to the Disco Fixed Record that will return
whether decode() has been called on it. This is now used to determine
whether various objects are ready. If we're received an update and it
decoded successfully for all fields we want, then an HLA object is
loaded and we can start tossing out PDUs.