Fix race condition during RPR mapper initialization.
When adding support for custom mappers we had to move pubsub into the
mapper itself. Previously this was done as part of connection init. What
resulted was a situation where at the time of subscription, a mapper was
potentially NOT subscribed to the HLA event bus. This caused it to miss
discovery calls and subsequently not see objects that entered a
federation before it.

The fix was to split when we subscribe to the HLA and DIS event busses.
First, we subscribe to the HLA event bus so we're ready for callbacks.
This is low risk because all mappers check that the incoming class of an
event is the one they're interested in, and there should never be a
match until subscription is complete.

Second, we tell the handler to initialize, which is where pubsub should
happen.

Finally, we get onto the DIS PDU/event bus. This has to come AFTER HLA
pubsub because otherwise we'd start translating PDUs and sending HLA
updates for them before we'd finished HLA publication.

Gut says there is still a smaller potential for issues here, but for now
it has solved the immediately reproducible issue.