Fix encode/decode errors in HLA when a crypto key/mode combo is used
In HLA/RPR a transmitter's crypto mode and key are stored in separate
fields, however in DIS they are both shoehorned into the cryptoKey field
using a bitmask.

Previously the DIS->HLA mapping code  was trying to populate the crypto
mode field with the crypto key, resulting in an invalid value exception
for the CryptographicModeEnum32 enum.

Likewise in the HLA->DIS code was setting the cryptoKey field twice,
once for the mode, and then overwriting it with the key.

Added the appropriate bit masking operations to extract and store these
values to/from the single DIS field

Fixes: CNR-1937