Fix decoding of modulation system detail for RPR RadioTransmitter.
This is a variant record in RPR and although the major type was being
set, the detail (which is in the body of the variant) was not being
updated and getting a default value of "0". This meant that in apps like
CNR that check that radio transmissio and receive properties are
properly aligned before allowing reception, HLA radios couldn't
communicate because the system detail setting was not surviving the
PDU->RPR conversion done by Disco.