Turn auto-provide switch on to deal with some end-user compat issues
DCS&S default FOM has made one change to the switches from the standard
RPR FOM that means there are issues that happen if a CNR federate is
responsible for creating the federation. This change aligns our default
RPR FOM with theirs. Re-judge this decision if others report
compatibility issues.