Fix the encoding type for spread sprectrum
According to the FOM, the spread spectrum enum should use
a uint16 value for the None value, but it was using a 32int instead,
which was causing decoder exceptions for some objects.
The None value is now using a uint16 to match the FOM, and the
discriminant type gets set based on frequency hopping.

Also a small fix for Mak rti using a modified name for privilege
to delete

Fixes: CNR-2020