Fix Mak RTI support when RTI provider is changed
Previously if the RTI provider was set to Mak after FOM moudles had
been initialised, it would cause errors due to the needed FOM files
not being extracted out of the jar, which Mak requires. Now a check
is done when the RTI provider is changed to Mak to make sure it can
load all of the FOM files.

Also fixes an NPE on federation cleanup if an RTI fails to start

Fixes: CNR-2026