Add check for EmitterBeam length > 255 to be compliant with DIS 7 spec.
DIS 7 Spec (7.6.2 paragraph f.5.i) says that for an emitter beam, if the
length of the beam is greater than 255, write it as 0.

I checked that the length for this record isn't being used anywhere in
deserialization (we go by the number of records, which is also written
to the record in the PDU), so this shouldn't have an advserse impact.