Update logging to stop overriding at root logger level (#61)
The previous update to logging was causing performing
operations on the root logger, which in turn could cause
issues when Disco was used in other applications that also
used Log4j, as it could result in their configuration being
disabled of unexpected manipulated. The code present is
now more in line with previous behaviour, but without the
deprecated method access.