Gets HLA working again with Java 21 (#67)
Fix ClassCastException when using Java 9+ and HLA.

For HLA support we extend the classpath at runtime. This is done by getting the system classloader and calling a method to add to the set of URLs it searches. This relies on the system classloader being a `URLClassLoader`. As of Java 9 it is not, and the reflection hack we use to make the private method that extends the classpath public so that we can call it is also no longer available (without pain). 

This update creates a custom class loader that we can extend the path on, chains it to the existing context class loader for the THREAD (no longer using the system classloader because it's not accessible), and then installs it as the threads context class loader. 

A little more limited (cannot extend the application classpath as a whole) but it should work.