Update to disco to be more compliant with FOM module loading
Disco was previously throwing an exception if two FOM modules defined the same
parameter for a class. However according to the spec, this is fine, as long as the
definition is the same. This now allows duplicate parameters.

NOTE: currently assuming that parameters with the same name ARE the same, which is
not necessarily the case.