Add support to RPR mappers for CPSM and SATCOM*
While CPSM does seem to be in the RPR 2.0 specification for the
MajorRFModulationTypeEnum16 enumeration, SATCOM isn't (although it has
found its way into the EBV). At the moment this change just adds mapper
support which should never get exercised, but it does not change the FOM
from teh standard FOM, which would in turn create the potential for FOM
merging issues.