Add PCAP support to Duplicator replay
Adds in the Sombrero library as a dependency, which provides Java PCAP
parsing and processing capabilities. Uses this to add support to the
Duplicator for reading PCAP files as a source of PDU data for replay.

With this patch, if you pass a .pcapng file on the command line to the
Duplicator, it will attempt to use the PCAP reader rather than the
default Duplicator session reader.

WARNING: This also updates the Eclipse project and Ant build to expect
JDK v21 due to requirements stemming from the Sombrero code.

Working: #72