Add PCAP support to Duplicator replay (#81)
* Add PCAP support to Duplicator replay

Adds in the Sombrero library as a dependency, which provides Java PCAP
parsing and processing capabilities. Uses this to add support to the
Duplicator for reading PCAP files as a source of PDU data for replay.

With this patch, if you pass a .pcapng file on the command line to the
Duplicator, it will attempt to use the PCAP reader rather than the
default Duplicator session reader.

WARNING: This also updates the Eclipse project and Ant build to expect
JDK v21 due to requirements stemming from the Sombrero code.

Working: #72

* Update build to packaging Java 21 JRE with installers

We previously supported Java 8. We are now updating to Java 21. With
this change the structure of the JDK has also changed. These updates
allow us to verify the presence of valid JDKs, and then package these
JDKs properly in the zip installers.

Working: #72