Add support for a FOM module override directory to RPR connection (#84)
This adds support to the RPR configuration for specifying a directory
that can be used to load modules from in place of the default modules
used inside the JAR file.

It is important to note that if you use this option, you will need to
specify/provide _all_ of the RPR modules, as the defaults will not be
used. This does however allow you to change the content of the modules.

Any .xml files inside the specified override directory will be treated
as modules and loaded.

Fix: #83