Fixes emitter beams not being cleared
Stored emitter beams were meant to be reset every time a new packet is
received, but an erroroneous call was preventing this. This commit fixes
this, making the emitter beam reset as intended.

Note: While this reset was clearly originally intended, it is technically
against the spec - Complete-Beam (see `EmitterStore.receivePdu`) allows for
only a subset of the beams to be in any given PDU. This reset would mean
that only the beams in the latest PDU with a partial list would exist,
resetting any beams recorded from previous partial PDUs. For now this
commit acts as a hotfix to beam set never being reset, at least for
testing purposes.