Loads DisApplication config before services
DisApplication services and helpers used to be loaded before the config
for the DisApplication was loaded. If they relied on config options then
they would only be initialised with the defaults, even if the config
used to initialise the DisApplication had been modified. This commit
reorders the DisApplication constructor to load the config first before
starting the services and helpers.